requirements.txt
fastapi
sqlalchemy
icecream
python-dotenv
uvicorn
gunicorn ; platform_system!="Windows"
uvloop ; platform_system!="Windows"
asyncpg
pydantic
pydantic[email]
pydantic-settings
aio-pika
git+https://github.com/Siva-Rajan-R/hyperlocal_platform